wordsAPI_vs_docker/requirements.txt
django
djangorestframework
django-cors-headers
sklearn
numpy 
pandas 
nltk
ufal.udpipe
requests
